requirements.txt
numpy
scipy
pynrrd
mrcfile
vtk
networkx
open3d
scikit-learn
scikit-image